Move `RawStr` to library
