+++ crates/smart-config/src/pat.rs
@@ -25,8 +25,57 @@ impl fmt::Display for PatternDisplay {
     fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Self::Exact(s) => write!(formatter, "{s:?}"),
-            Self::Regex(regex) => write!(formatter, "Regex({regex:?})"),
+            Self::Regex(regex) => write!(formatter, "Regex({})", RawStr(regex)),
             Self::Generic(s) => formatter.write_str(s),
+        }
+    }
+}
+
+/// Wrapper for strings that outputs a string as a raw string literal, like `r"\s+"`.
+#[doc(hidden)] // reused in the `commands` crate; logically private
+#[derive(Clone, Copy)]
+pub struct RawStr<'a>(pub &'a str);
+
+impl fmt::Debug for RawStr<'_> {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::Display::fmt(self, formatter)
+    }
+}
+
+impl fmt::Display for RawStr<'_> {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let hash_count = self.hash_count();
+        write!(formatter, "r")?;
+        for _ in 0..hash_count {
+            write!(formatter, "#")?;
+        }
+        write!(formatter, "\"{}\"", self.0)?;
+        for _ in 0..hash_count {
+            write!(formatter, "#")?;
+        }
+        Ok(())
+    }
+}
+
+impl RawStr<'_> {
+    // Determine the number of necessary `#` for the raw string specifier.
+    fn hash_count(self) -> usize {
+        let has_double_quotes = self.0.chars().any(|ch| ch == '"');
+        if has_double_quotes {
+            let mut max_hashes = 0;
+            let mut hash_start = None;
+            for (i, ch) in self.0.chars().enumerate() {
+                if ch == '#' {
+                    if hash_start.is_none() {
+                        hash_start = Some(i);
+                    }
+                } else if let Some(hash_start) = hash_start.take() {
+                    max_hashes = max_hashes.max(i - hash_start);
+                }
+            }
+            max_hashes + 1
+        } else {
+            0
         }
     }
 }
@@ -86,5 +135,29 @@ impl Split for &'static LazyLock<Regex> {
 
     fn display(&self) -> PatternDisplay {
         PatternDisplay::Regex(self.as_str().to_owned())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn hash_count_for_raw_strings_is_correct() {
+        let s = RawStr("Hello, world!");
+        assert_eq!(s.hash_count(), 0);
+        assert_eq!(s.to_string(), "r\"Hello, world!\"");
+
+        let s = RawStr("####");
+        assert_eq!(RawStr("####").hash_count(), 0);
+        assert_eq!(s.to_string(), "r\"####\"");
+
+        let s = RawStr(r#"x="1""#);
+        assert_eq!(s.hash_count(), 1);
+        assert_eq!(s.to_string(), "r#\"x=\"1\"\"#");
+
+        let s = RawStr(r##"x="#1""##);
+        assert_eq!(s.hash_count(), 2);
+        assert_eq!(s.to_string(), "r##\"x=\"#1\"\"##");
     }
 }
